MDEXP-641 Refactored tests
